+++ frontend/src/game/Tournament/index.js
@@ -1,2 +1,3 @@
 export { default as TournamentTable } from "./tournamentTable.js";
 export { default as TournamentResult } from "./tournamentResult.js";
+export { default as TournamentTeam} from "./tournamentTeam.js";
